rentally_backend/requirements.txt
# Django & REST Framework
Django==5.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Storage & Media
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

# Password Hashing
bcrypt==4.1.0

# Static Files
whitenoise==6.6.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1

# Validation
django-extensions==3.2.3
django-filter==23.5

# Monitoring & Logging
sentry-sdk==1.35.0

# Testing
pytest==7.4.0
pytest-django==4.7.0
factory-boy==3.3.0

# Real-time / WebSockets
daphne==4.1.2
channels==4.1.0
